Restructure and add rust like conversion to and from
Fixes #30
